Implement N50 and modify Reporter interface to reflect that an int32 is returned rather than a float32

+++ reporter/interface.go
@@ -5,7 +5,7 @@ type Reporter interface {
 	// Process constructs the report for the given assembler results
 	Process() error
 	// getL50 computes and returns the L50 score of the report contigs
-	getL50() (float32, error)
+	getL50() (int32, error)
 	// getN50 computes and returns the N50 score of the report contigs
-	getN50() (float32, error)
+	getN50() (int32, error)
 }
