+++ reporter/l50.go
@@ -5,6 +5,6 @@ import (
 )
 
 // getL50 computes and returns the L50 score of the report contigs
-func (r *report) getL50() (float32, error) {
+func (r *report) getL50() (int32, error) {
 	return .0, fmt.Errorf("getL50 not implemented yet")
 }
